refactor(agent): extract SWM host-mode subsystem into mixin holder
Moves the SWM host-mode cluster (~2.6k lines) out of dkg-agent.ts into
SwmHostModeMethods (extends DKGAgentBase): host-mode store init/reconcile/
wire, envelope + ciphertext-chunk ingest, host-catchup + get-chunk handlers,
chain-ordinal VM reconcile, and the catchup/enable/stats entrypoints.
Assembled onto DKGAgent via applyMixins + interface merge.

- moved methods take `this: DKGAgent` and are public (cross-called from main)
- hostModePersistenceQueues field was declared mid-cluster; relocated to
  DKGAgentBase (applyMixins copies methods only, not field initializers)
- 10 sibling methods the cluster calls widened private->public
- one DKGAgent.VM_RECONCILE_CONFIRMATION_DEPTH static ref -> DKGAgentBase

Pure relocation. dkg-agent.ts 19884 -> 17245. Typecheck clean; unit (123)
+ host-mode/chunk/reconcile suites (48) + agent.test.ts (126) green.
